requirements.txt
Pillow==2.4.0
bottle==0.12.5
wsgiref==0.1.2
